finish exercise nove !
