+++ test/hw6/3_exercise.py
@@ -1,0 +1,13 @@
+
+
+
+def is_prime(number):
+    if number < 2:
+        return False
+    for i in range(2, int(number ** 0.5) + 1):
+        if number % i == 0:
+            return False
+    return True
+print(is_prime(11))
+print(is_prime(121))
+
