+++ test/hw6/6_exercise.py
@@ -1,0 +1,7 @@
+
+
+
+def calculate_sum(*args):
+    return sum(args)
+print(calculate_sum(1, 2, 3, 4, 5))
+
