+++ test/hw6/7_exercise.py
@@ -1,0 +1,9 @@
+
+
+
+def remove_vowels(input_string):
+    vowels = "aeiouAEIOU"
+    return ''.join(char for char in input_string if char not in vowels)
+print(remove_vowels("Hello, World!"))
+
+
